Made sort an array to handle multiple options
