+++ src/lib/db/review/index.ts
@@ -5,69 +5,17 @@ import {
   RPC_GET_DISTINCT_REVIEW_IMAGES,
   RPC_GET_PRODUCT_REVIEWS_SUMMARY,
 } from '../constants/databaseTableNames';
-import { Tables } from '../types';
 import { getPagination } from '../utils';
 import { supabaseDatabaseClient } from '../supabaseClients';
-
-export type TReviewData = Tables<'reviews-2'>;
-
-export type TProductReviewsQueryFilters = {
-  productType?: 'Car Covers' | 'SUV Covers' | 'Truck Covers' | 'Seat Covers';
-  year?: string;
-  make?: string;
-  model?: string;
-};
-
-export type FilterParams = {
-  field: string;
-  operator: 'eq' | 'neq' | 'gt' | 'lt' | 'gte' | 'lte' | 'notnull';
-  value: string | number;
-};
-
-export type SortParams = {
-  field: string;
-  order: 'asc' | 'desc';
-};
-
-export type TProductReviewsQueryOptions = {
-  pagination?: {
-    page?: number;
-    limit?: number;
-  };
-  sort?: SortParams;
-  filters?: FilterParams[];
-  search?: string;
-};
-
-export type TProductReviewSummary = {
-  total_reviews: number;
-  average_score: number;
-};
-
-export type TProductReviewDistinctImages = {
-  rating_stars: number;
-  helpful: number;
-  reviewed_at: string;
-  gpt_review_id: string;
-  model: string;
-  year_generation: string;
-  submodel1: string;
-  submodel2: string;
-  mirror: string;
-  review_description: string;
-  make_slug: string;
-  review_title: string;
-  review_author: string;
-  review_image: string;
-  model_slug: string;
-  size: string;
-  sku: string;
-  parent_generation: string;
-  product_type: string;
-  product_name: string;
-  type: string;
-  make: string;
-};
+import { TReviewData } from '..';
+import {
+  TProductReviewsQueryFilters,
+  TProductReviewsQueryOptions,
+  ProductReviewsQueryFiltersSchema,
+  ProductReviewsQueryOptionsSchema,
+  TProductReviewDistinctImages,
+  TProductReviewSummary,
+} from './types';
 
 export const generateSlug = (text: string) => {
   if (!text) return ''; // Return an empty string if text is falsy
@@ -122,45 +70,6 @@ const retry = async <T>(
   throw new Error('Unexpected error in retry logic.');
 };
 
-const FilterSchema = z.object({
-  field: z.string(),
-  operator: z.enum(['eq', 'neq', 'gt', 'lt', 'gte', 'lte', 'notnull']),
-  value: z.union([z.string(), z.number()]), // Adjust as necessary
-});
-
-const ProductReviewsQueryFiltersSchema = z.object({
-  productType: z.optional(
-    z.union([
-      z.literal('Car Covers'),
-      z.literal('SUV Covers'),
-      z.literal('Truck Covers'),
-      z.literal('Seat Covers'),
-    ])
-  ),
-  year: z.string().optional(),
-  make: z.string().optional(),
-  model: z.string().optional(),
-});
-
-const ProductReviewsQueryOptionsSchema = z.object({
-  pagination: z
-    .object({
-      page: z.number().optional().default(1),
-      limit: z.number().optional().default(8),
-    })
-    .optional()
-    .default({ page: 1, limit: 8 }),
-  sort: z
-    .object({
-      field: z.string().default('helpful'),
-      order: z.enum(['asc', 'desc']).default('desc'),
-    })
-    .optional()
-    .default({ field: 'helpful', order: 'desc' }),
-  filters: z.array(FilterSchema).optional(),
-  search: z.string().optional(),
-});
-
 export async function getProductReviewsByPage(
   productQueryFilters: TProductReviewsQueryFilters,
   options: TProductReviewsQueryOptions,
@@ -189,9 +98,7 @@ export async function getProductReviewsByPage(
         .select(
           'review_image,review_description,review_title,rating_stars,review_author,helpful,reviewed_at'
         )
-        .not('sku', 'like', '%N/A%')
         .not('review_author', 'is', null)
-
         .range(from, to);
 
       if (productType) {
@@ -240,13 +147,17 @@ export async function getProductReviewsByPage(
         }
       });
 
-      if (productType === 'Seat Covers') {
-        fetch = fetch.neq('sku', 'CL-SC-10--BK-1TO-')
-        fetch = fetch.order('sku', { ascending: true });
-        fetch = fetch.order('helpful', { nullsFirst: false, ascending: false });
-      } else if (sort && sort.field) {
-        fetch = fetch.order(sort.field, { ascending: sort.order === 'asc' });
-      }
+      // if (productType === 'Seat Covers') {
+      //   fetch = fetch.neq('sku', 'CL-SC-10--BK-1TO-')
+      //   fetch = fetch.order('sku', { ascending: true });
+      //   fetch = fetch.order('helpful', { nullsFirst: false, ascending: false });
+      // } else if (sort && sort.field) {
+      //   fetch = fetch.order(sort.field, { ascending: sort.order === 'asc' });
+      // }
+
+      sort.forEach(({ field, order, nullsFirst }) => {
+        fetch = fetch.order(field, { ascending: order === 'asc', nullsFirst });
+      });
 
       // if (search) {
       //   fetch = fetch.textSearch('review_description', search);
@@ -491,9 +402,9 @@ export async function getProductReviewsByImage(
       }
     });
 
-    if (sort && sort.field) {
-      fetch = fetch.order(sort.field, { ascending: sort.order === 'asc' });
-    }
+    sort.forEach(({ field, order, nullsFirst }) => {
+      fetch = fetch.order(field, { ascending: order === 'asc', nullsFirst });
+    });
 
     const { data, error } = await fetch;
